Move api method => endpoint method
